+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed unpacked fields not accepting packed bytes
+- Unknown fields are now preserved and roundtripped.
 
 ## [0.3.5] - 2022-01-02